# Sequential lda optimization