… gensim/models/ldaseq_sslm_inner.pyx
… gensim/models/ldaseqmodel_optimized.py