Hi team,

This pull request addresses the issue of the slow computational time of LdaSeqModel #1545 . In the current stage, it contains cythonized methods of SSLM and LdaPost classes. It is not the final version, I plan to add the following changes:
- cythonize ```sslm_counts_init``` function of sslm class 
- refactor code (e.g. ```chain.sslm_counts_init(topic_obs_variance, topic_chain_variance, sstats)``` in ```LdaSeqModel``` class, here class method is used, however I don't know why) + many others 
- add more documentation for functions
- implement DIM model 
- tune memory usage and allocation 
- fix bug which I found out in the ```update_obs``` function of ```sslm``` class. You can compare the logic in the gensim [implementation](https://github.com/RaRe-Technologies/gensim/blob/57b3af31c2905fa9e3abe6c1002212b17e5c5b3d/gensim/models/ldaseqmodel.py#L1089) and Blei's lab [implementation](https://github.com/blei-lab/dtm/blob/810d9abb5eda13a486f3fb283383981a2d7aeb66/dtm/ss-lm.c#L1148)

By now the performance improvement is 3-5 times on toy datasets e.g. https://github.com/RaRe-Technologies/gensim/blob/develop/gensim/test/test_ldaseqmodel.py. I still haven't tried to benchmark for larger datasets as it takes a while to train original implementation, however, the speed up seems to be much higher based on synthetic data I tried.  
 
Extracted topics of optimized model differ a bit from the original one (gensim tests are passed). The exact places in the code which causes it are
- ```update_zeta``` function of ```sslm``` class 
- ```np.negative``` at the end of ```df_obs``` 
Everything except these two places works absolutely the same. From my perspective, it is not a bug in the code, just issues with the precision. (I hope so)

@gojomo @mpenkov What do you think about the idea of such changes in gensim in general, code structure and code itself particularly? I plan to finish with the optimization and bug fixes I mentioned above by the end of this month. 

P.S. I've created the pull request before ending to be sure that these changes are valuable for the project :)