@Sharganov 2021-06-23
I've run a test using setup from ldaseqmodel example [notebook](https://github.com/RaRe-Technologies/gensim/blob/develop/docs/notebooks/ldaseqmodel.ipynb) in gensim docs. As for now, the current time distribution is present on the picture below. Almost all time is spent in scipy Python optimization code. The time which is spent in Cython marked with red rectangles.

The test was done using Intel(R) Core(TM) i7-4720HQ CPU @ 2.60GHz

![ldaseq_optimized_news](https://user-images.githubusercontent.com/8983727/123161012-8e634800-d477-11eb-8dee-38b630a63740.PNG)


The raw output of a profiler + file for kcachegrind is on [gdrive](https://drive.google.com/drive/folders/1tEDaKrWmrQ0HgGIdrdsl76XwJngVh0-O?usp=sharing)