@Sharganov 2021-07-03
I trained both implementations using the first 10 years of the [UN General Debate dataset ](https://www.kaggle.com/unitednations/un-general-debates/home). The code I used:
```
from gensim.corpora import Dictionary
from gensim import utils
from gensim.models import LdaSeqModel

import numpy as np
import pandas as pd

data = pd.read_csv("datasets/un-general-debates.csv")
data["year"] = data["year"].apply(lambda x: int(x))
data = data[data["year"] < 1980]

paragraphs = [[(p,year) for p in speech.split("\n") if len(p) > 10] for speech, year  in zip(data.text, data.year)]
paragraphs = pd.DataFrame(data=sorted([p for s in paragraphs for p in s], key=lambda x: x[1]), columns=["text", "year"])

time_slices = paragraphs.year.value_counts().values
paragraphs = [utils.simple_preprocess(p) for p in paragraphs.text]

dictionary = Dictionary(paragraphs)
corpus = [dictionary.doc2bow(text) for text in paragraphs]

del data, paragraphs

model = LdaSeqModel(corpus=corpus,id2word=dictionary, num_topics=15, time_slice=time_slices)

```

The results:
  Number of docs: 61674
  Vocabulary Length: 29485
  Number of topics: 15
  Number of time slices: 10

  **Old Implementation** : ~47 hours
  **New Implementation** ~3 hours 40 mins