@piskvorky 2021-07-03
Thanks for the timings. And are the new-vs-old results comparable, quality-wise?