@Sharganov 2021-07-04
Sure... I accidentally rewrote a trained old model :( 

So for now I just trained with two years of data using the same code + removed stopwords using nltk set. I've put the resulting topics in the excel [table](https://docs.google.com/spreadsheets/d/1VV12tw7PMRGuc_66da9WsWzcc8pfPhMbZhrjUyqMTo0/edit?usp=sharing) . As for me topics are very similar. 

Regarding the model trained with 10 years of data, let me train it one more time :) I'll back in a couple of days. Meanwhile, I'll find a more formal way to compare two SeqLda models.