@Sharganov 2021-07-07
I've trained the model one more time. Pickled models are available on [gdrive](https://drive.google.com/drive/folders/1tEDaKrWmrQ0HgGIdrdsl76XwJngVh0-O?usp=sharing). I calculated the coherence score for both models as it was done in the docs [tutorial](https://github.com/RaRe-Technologies/gensim/blob/develop/docs/notebooks/ldaseqmodel.ipynb).
I am not sure that it is the right choice to compare classic vs optimized models using a coherence score, I think they should be equal in a perfect case. Nevertheless, they are very close for all time slices. 

The code I used: 
```
from gensim.models.coherencemodel import CoherenceModel
import pickle

[CoherenceModel(topics=model.dtm_coherence(time=i), corpus=corpus, 
                dictionary=dictionary, coherence='u_mass').get_coherence() 
 for i in range(len(time_slices))]
```
**Results:**
| Time slice        | Old           | New  |
| ------- |:-------------:|:-----:|
|       1  | -2.390386933137473 |  -2.3849280478010373|
|       2  | -2.3854065916265434      |  -2.38503617012231 |
|   3       | -2.412550400281885      |    -2.412624753824853 |
|   4       | -2.403873956103539     |    -2.402234764724972 |
|   5       |  -2.4228081694247994      |    -2.416472944689486 |
|  6      |  -2.462690192130912      |     -2.466086114010153 |
| 7       | -2.4189737189539      |    -2.4169199318713814, |
| 8       |  -2.484822837691009      |    -2.488989346989372 |
| 9       | -2.5564440634792547      |    -2.557404323965962|
| 10       |  -2.551624889028128     |    -2.54892377159758] |