@piskvorky 2021-09-01
Good point. @Sharganov can you fix the bug you found above & tune memory? Let's get this merged & released!