@Sharganov 2022-01-15
Hi everyone, sorry for a long break. I fixed the bug which I mentioned. Also I'll look at memory usage.