@piskvorky 2022-02-19
@Sharganov what did you find about the memory? We're planning a new release, and this looks a solid candidate for including if finished up.