@piskvorky 2022-02-22
No problem, thanks. If we don't the optimization into this release, we can include it in the next.