@piskvorky 2022-06-16
@Sharganov how did your memory optimization go? Looks like there is quite some demand for this feature!